refactor(research): extract mission lifecycle
